chore: 폴더 이동 (common => DetailPage)
- Mypage와 DetailPage 통합
