+++ client/src/common/index.ts
@@ -3,6 +3,5 @@ export { CommonLayout } from './CommonLayout';
 export { MiniNavBar } from './MiniNavBar';
 export { InterestInput, TechStackInput } from './DropdownInput';
 export { Button } from './Button';
-export { ViewModeContainer } from './ViewModeContainer';
 export { CodeBox } from './CodeBox';
 export { NavSubtitle } from './NavSubtitle';
