+++ client/src/pages/DetailPage/EditModeContainer/BottomProfileEditor.styles.ts
