+++ client/src/pages/DetailPage/EditModeContainer/BottomProfileEditor.tsx
