+++ client/src/pages/DetailPage/EditModeContainer/CodeEditor.styles.ts
