+++ client/src/pages/DetailPage/ViewModeContainer/BottomProfileBox.styles.ts
