+++ client/src/pages/DetailPage/ViewModeContainer/BottomProfileBox.tsx
