+++ client/src/pages/DetailPage/ViewModeContainer/InterestBadge/constants.ts
