+++ client/src/pages/DetailPage/ViewModeContainer/InterestBadge/index.tsx
