+++ client/src/pages/DetailPage/ViewModeContainer/TechStackBadge/index.tsx
