+++ client/src/pages/DetailPage/ViewModeContainer/TechStackBadge/styles.ts
