+++ client/src/pages/DetailPage/ViewModeContainer/TopProfileBox.styles.ts
