+++ client/src/pages/DetailPage/ViewModeContainer/TopProfileBox.tsx
