+++ client/src/pages/DetailPage/ViewModeContainer/index.tsx
