+++ client/src/pages/DetailPage/ViewModeContainer/styles.ts
