+++ client/src/pages/DetailPage/styles.ts
