+++ client/src/pages/index.ts
@@ -1,6 +1,5 @@
 export { MainPage } from './MainPage';
 export { DetailPage } from './DetailPage';
-export { Mypage } from './Mypage';
 export { SettingsPage } from './SettingsPage';
 
 export { LoginPage } from './LoginPage';
